requirements.txt
PySide6>=6.5.0
pywin32>=306
pdf2docx>=0.5.8
pdf2image>=1.17.0
Pillow>=10.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
PyMuPDF>=1.24.0
